+++ kit/feature/list.go
@@ -86,34 +86,6 @@ func QueryTracing() BoolFlag {
 	return queryTracing
 }
 
-var notebooks = MakeBoolFlag(
-	"Notebooks",
-	"notebooks",
-	"Monitoring Team",
-	false,
-	Temporary,
-	true,
-)
-
-// Notebooks - Determine if the notebook feature's route and navbar icon are visible to the user
-func Notebooks() BoolFlag {
-	return notebooks
-}
-
-var notebooksApi = MakeBoolFlag(
-	"Notebooks Service API",
-	"notebooksApi",
-	"Edge Team",
-	true,
-	Temporary,
-	true,
-)
-
-// NotebooksServiceApi - Enable the Equivalent notebooksd Service API
-func NotebooksServiceApi() BoolFlag {
-	return notebooksApi
-}
-
 var injectLatestSuccessTime = MakeBoolFlag(
 	"Inject Latest Success Time",
 	"injectLatestSuccessTime",
@@ -154,34 +126,6 @@ var timeFilterFlags = MakeBoolFlag(
 // TimeFilterFlags - Filter task run list based on before and after flags
 func TimeFilterFlags() BoolFlag {
 	return timeFilterFlags
-}
-
-var csvUploader = MakeBoolFlag(
-	"UI CSV Uploader",
-	"csvUploader",
-	"Monitoring Team",
-	true,
-	Temporary,
-	true,
-)
-
-// UiCsvUploader - Adds the ability to upload data from a CSV file to a bucket
-func UiCsvUploader() BoolFlag {
-	return csvUploader
-}
-
-var editTelegrafs = MakeBoolFlag(
-	"Editable Telegraf Configurations",
-	"editTelegrafs",
-	"Monitoring Team",
-	true,
-	Temporary,
-	true,
-)
-
-// EditableTelegrafConfigurations - Edit telegraf configurations from the UI
-func EditableTelegrafConfigurations() BoolFlag {
-	return editTelegrafs
 }
 
 var cursorAtEOF = MakeBoolFlag(
@@ -230,7 +174,7 @@ var annotations = MakeBoolFlag(
 	"Annotations UI",
 	"annotations",
 	"Monitoring Team",
-	false,
+	true,
 	Temporary,
 	true,
 )
@@ -247,13 +191,9 @@ var all = []Flag{
 	memoryOptimizedFill,
 	memoryOptimizedSchemaMutation,
 	queryTracing,
-	notebooks,
-	notebooksApi,
 	injectLatestSuccessTime,
 	enforceOrgDashboardLimits,
 	timeFilterFlags,
-	csvUploader,
-	editTelegrafs,
 	cursorAtEOF,
 	refreshSingleCell,
 	rangeAnnotations,
@@ -267,13 +207,9 @@ var byKey = map[string]Flag{
 	"memoryOptimizedFill":           memoryOptimizedFill,
 	"memoryOptimizedSchemaMutation": memoryOptimizedSchemaMutation,
 	"queryTracing":                  queryTracing,
-	"notebooks":                     notebooks,
-	"notebooksApi":                  notebooksApi,
 	"injectLatestSuccessTime":       injectLatestSuccessTime,
 	"enforceOrgDashboardLimits":     enforceOrgDashboardLimits,
 	"timeFilterFlags":               timeFilterFlags,
-	"csvUploader":                   csvUploader,
-	"editTelegrafs":                 editTelegrafs,
 	"cursorAtEOF":                   cursorAtEOF,
 	"refreshSingleCell":             refreshSingleCell,
 	"rangeAnnotations":              rangeAnnotations,