+++ CHANGELOG.md
@@ -22,6 +22,10 @@ This release upgrades the project to `go` version 1.16.
 
 Because of the version bump to `go`, the macOS build for this release requires at least version 10.12 Sierra to run.
 
+### Notebooks and Annotations
+
+Support for Notebooks and Annotations is included with this release.
+
 ### SQLite Metadata Store
 
 This release adds an embedded SQLite database for storing metadata required by the latest UI features like Notebooks and Annotations.
@@ -53,6 +57,7 @@ This release adds an embedded SQLite database for storing metadata required by t
 1. [21910](https://github.com/influxdata/influxdb/pull/21910): Added `--ui-disabled` option to `influxd` to allow for running with the UI disabled.
 1. [21938](https://github.com/influxdata/influxdb/pull/21938): Added route to delete individual secret.
 1. [21958](https://github.com/influxdata/influxdb/pull/21958): Telemetry improvements: Do not record telemetry data for non-existant paths; replace invalid static asset paths with a slug.
+1. [21972](https://github.com/influxdata/influxdb/pull/21972): Added support for notebooks and annotations.
 
 ### Bug Fixes